+++ HISTORY.md
@@ -1,6 +1,9 @@
 History
 ===============
 
+### By now 2015-04-01
+* Add -v option for bin/biotcm
+
 ### v0.2.2 2015-03-18
 * Fix a bug in BioTCM::Databases::OMIM